feat: make medication scheduler timezone-aware, execute transition plan steps
- Load user TZ from timezone_history; fall back to time.Local if not set
- Compute daily/weekly dose targets in user timezone rather than system TZ
- When an APPROVED transition plan exists, use plan step ScheduledAt instead
  of normal schedule; mark each step consumed after CreateIntake succeeds
- Skip normal scheduling for medications that have pending plan steps
  (even if the next step is not yet due), preventing schedule overlap
- Idempotent: if an intake already exists at a step's time, just mark the
  step consumed without creating a duplicate intake
- Extend MedicationStore interface with GetCurrentTimezone,
  GetLatestActiveOrPendingTZTransitionPlan, GetPendingStepsForPlan,
  MarkStepConsumed
- Add medication_tz_test.go with 6 scenarios: user-TZ scheduling fires/not,
  approved-plan step used, future step suppresses normal schedule, partial
  consumption resumes remaining steps, cancelled plan → normal schedule,
  idempotent re-run of already-scheduled step
- Add optional user_timezone field to JSON test scenario format to decouple
  tests from system timezone
- Update CLAUDE.md: remove deferred caveat, document tz_transition_plans and
  tz_transition_steps tables

+++ internal/domain/tzreschedule/policy.go
@@ -41,8 +41,8 @@ func MaxShiftPerDose(p Policy) time.Duration {
 // MinDoseInterval returns the minimum allowed gap between two consecutive doses
 // for a medication with the given nominal schedule interval (in hours).
 // This is a hard constraint that must never be violated.
-func MinDoseInterval(scheduleIntervalHours int, p Policy) time.Duration {
-	base := time.Duration(scheduleIntervalHours) * time.Hour
+func MinDoseInterval(scheduleIntervalHours float64, p Policy) time.Duration {
+	base := time.Duration(scheduleIntervalHours * float64(time.Hour))
 	switch p {
 	case PolicyStrict:
 		return time.Duration(float64(base) * 0.70)
@@ -56,8 +56,8 @@ func MinDoseInterval(scheduleIntervalHours int, p Policy) time.Duration {
 // MaxDoseInterval returns the maximum allowed gap between two consecutive doses
 // for a medication with the given nominal schedule interval (in hours).
 // This is a hard constraint that must never be violated.
-func MaxDoseInterval(scheduleIntervalHours int, p Policy) time.Duration {
-	base := time.Duration(scheduleIntervalHours) * time.Hour
+func MaxDoseInterval(scheduleIntervalHours float64, p Policy) time.Duration {
+	base := time.Duration(scheduleIntervalHours * float64(time.Hour))
 	switch p {
 	case PolicyStrict:
 		return time.Duration(float64(base) * 1.50)
